requirements.txt
pytorch
typer[all]
pydantic
loguru
